backend/liftbot_server/requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
openai==1.61.1
python-dotenv==1.0.1
Requests==2.32.3
